Removido tudo de Unidade, Army integrado
- Algumas questões de GUI foram resolvidas, outras ainda estão
pendentes.
- Deletado Mundo e colocado ItemPaldino no local adequado

+++ src/io/github/avatarhurden/tribalwarsengine/ferramentas/dados_de_unidade/InformaçãoTropaPanel.java
@@ -187,9 +187,9 @@ public class InformaçãoTropaPanel {
 		}
 		
 		danoLabel.setText(getFormattedNumber(army.getAtaque()));
-		defGeralLabel.setText(getFormattedNumber(army.getDefesaGeral()));
-		defCavaloLabel.setText(getFormattedNumber(army.getDefesaCavalaria()));
-		defArqueiroLabel.setText(getFormattedNumber(army.getDefesaArqueiro()));
+		defGeralLabel.setText(getFormattedNumber((int) army.getDefesaGeral()));
+		defCavaloLabel.setText(getFormattedNumber((int) army.getDefesaCavalaria()));
+		defArqueiroLabel.setText(getFormattedNumber((int) army.getDefesaArqueiro()));
 
 		saqueLabel.setText(getFormattedNumber(army.getSaque()));
